Move Device & co to device/_main.py
